+++ tests/test_utilities/test_csvsort.py
@@ -41,6 +41,16 @@ class TestCSVSort(CSVKitTestCase, ColumnsTests, EmptyFileTests, NamesTests):
         new_order = [str(r[0]) for r in reader]
         self.assertEqual(test_order, new_order)
 
+    def test_ignore_case(self):
+        self.assertRows(['-i', 'examples/test_ignore_case.csv'], [
+            ['a', 'b', 'c'],
+            ['3', '2009-01-01', 'd'],
+            ['20', '2001-01-01', 'c'],
+            ['20', '2002-01-01', 'b'],
+            ['100', '2003-01-01', 'a'],
+            ['100', '2003-01-01', 'A'],
+        ])
+
     def test_no_blanks(self):
         reader = self.get_output_as_reader(['examples/blanks.csv'])
         test_order = [