+++ CHANGELOG.rst
@@ -3,6 +3,7 @@ Unreleased
 
 * feat: Add support for Zstandard files with the ``.zst`` extension, if the ``zstandard`` package is installed.
 * feat: :doc:`/scripts/csvformat` adds a :code:`--out-asv` (:code:`--A`) option to use the ASCII unit separator and record separator.
+* feat: :doc:`/scripts/csvsort` adds a :code:`--ignore-case` (:code:`--i`) option to perform case-independent sorting.
 
 1.4.0 - February 13, 2024
 -------------------------